pyproject.toml
[tool.poetry]
name = "Examiner"
version = "1.0.0"
description = ""
authors = ["Xewus <xewuss@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
mysqlclient = "^2.1.1"
Django = "^4.1"
uWSGI = "^2.0.20"
python-decouple = "^3.6"
daphne = "^3.0.2"
requests = "^2.28.1"
pytest-django = "^4.5.2"
django-extensions = "^3.2.0"
pytest-cov = "^3.0.0"
pytest-pythonpath = "^0.7.4"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"
pytest-django = "^4.5.2"
uwsgitop = "^0.11"
django-debug-toolbar ="3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
